Public Library

Formulation Function

Missing docstring.

Missing docstring for solve_restoration_full. Check Documenter's build log for details.

EGRIP.solve_restoration_part — Function

Solve partial restoration problem (The restoration problem could be partial or full restorations)

Partial restoration problem assumes that a part of the network is still functioning.

Inputs: A set of restoration data in csv format and original system data

network data directory where data format could be json, matpower and psse

restoration data directory

restoration_gen: specify initial generator status, cranking specifications and black-start generators

restoration_bus: specify initial bus status and its load priority, from where the problem type (partial or full restorations) can be determined

restoration_line: specify initial line status, from where the problem type (partial or full restorations) can be determined

result storage directory

gap

Output: Restoration plans

Constraints:

linearized AC power flow constraint

steady-state voltage variation constraint

generator cranking constraint

generator status and output constraint

load pick-up constraint

EGRIP.solve_section — Function

Solve sectionalization problem for restoration preparedness

Problem type: The sectionalization problem could assign all buses to certain sections or only critical buses.

Inputs:

network data directory

restoration data directory

result storage directory

gap

Output:

JSON file of network data of each section

Constraints:

Missing docstring.

Missing docstring for solve_startup. Check Documenter's build log for details.