Internal Library

Integrated Restoration

use @autodocs block to splice the docstrings

EGRIP.solve_restoration_full — Method

Solve restoration problem

Problem type: The restoration problem could be partial or full restorations

Inputs: A set of restoration data in csv format and original system data

network data directory where data format could be json, matpower and psse

restoration data directory

restoration_gen: specify initial generator status, cranking specifications and black-start generators

restoration_bus: specify initial bus status and its load priority, from where the problem type (partial or full restorations) can be determined

restoration_line: specify initial line status, from where the problem type (partial or full restorations) can be determined

result storage directory

gap

Output: Restoration plans

Constraints:

linearized AC power flow constraint

steady-state voltage variation constraint

generator cranking constraint

generator status and output constraint

load pick-up constraint

Sectionalization Models

use @autodocs block to splice the docstrings

EGRIP.solve_section — Method

Solve sectionalization problem for restoration preparedness

Problem type: The sectionalization problem could assign all buses to certain sections or only critical buses.

Inputs:

network data directory

restoration data directory

result storage directory

gap

Output:

JSON file of network data of each section

Constraints:

Power Flow Models

use @autodocs block to splice the docstrings

EGRIP.form_branch — Method

branch (power flow) constraints

linearized power flow

\[\begin{align*} p_{bij,t}=G_{ii}(2vl_{ij,t}-x_{ij,t}) + G_{ij}(vl_{ij,t} + vl_{ji,t}-x_{ij,t}) + B_{ij}(al_{ij,t}-al_{ij,t})\\ q_{bij,t}=-B_{ii}(2vl_{ij,t}-x_{ij,t}) - B_{ij}(vl_{ij,t} + vl_{ji,t}-x_{ij,t}) + G_{ij}(al_{ij,t}-al_{ij,t})\\ \end{align*}\]

EGRIP.form_nodal — Method

form the nodal constraints:

voltage constraint

voltage deviation should be limited

voltage constraints are only activated if the associated line is energized

\[\begin{align*} & v^{\min}_{i} \leq v_{i,t} \leq v^{\max}_{i}\\ & v^{\min}_{i}x_{ij,t} \leq vl_{ij,t} \leq v^{\max}_{i}x_{ij,t}\\ & v^{\min}_{j}x_{ij,t} \leq vl_{ji,t} \leq v^{\max}_{j}x_{ij,t}\\ & v_{i,t} - v^{\max}_{i}(1-x_{ij,t}) \leq vl_{ij,t} \leq v_{i,t} - v^{\min}_{i}(1-x_{ij,t})\\ & v_{j,t} - v^{\max}_{j}(1-x_{ij,t}) \leq vl_{ij,t} \leq v_{j,t} - v^{\min}_{j}(1-x_{ij,t}) \end{align*}\]

angle difference constraint

angle difference should be limited

angle difference constraints are only activated if the associated line is energized

\[ \begin{align*} & a^{\min}_{ij} \leq a_{i,t}-a_{j,t} \leq a^{\max}_{ij}\\ & a^{\min}_{ij}x_{ij,t} \leq al_{ij,t}-al_{ji,t} \leq a^{\max}_{ij}x_{ij,t}\\ & a_{i,t}-a_{j,t}-a^{\max}_{ij}(1-x_{ij,t}) \leq al_{ij,t}-al_{ji,t} \leq a_{i,t}-a_{j,t}-a^{\min}_{ij}(1-x_{ij,t}) \end{align*}\]

generator and bus energizing logics

energized line cannot be shut down

bus should be energized before the connected genertor being on

\[\begin{align*} & x_{ij,t} \geq x_{ij,t-1}\\ & u_{i,t} \geq x_{ij,t}\\ & u_{j,t} \geq x_{ij,t} \end{align*}\]

bus energized constraints

bus energized indicating generator energized

energized buses cannot be shut down

\[\begin{align*} & v^{\min}u_{i,t} \leq vb_{i,t} \leq v^{\max}u_{i,t} \\ & v_{i,t} - v^{\max}(1-u_{i,t}) \leq vb_{i,t} \leq v_{i,t} - v^{\min}(1-u_{i,t})\\ & u_{g,t} = y_{g,t}\\ & u_{i,t} \geq u_{i,t-1} \end{align*}\]

nodal power balance constraint

\[\begin{align*} & \sum_{b\in i}p_{b,t}=\sum_{g\in i}pg_{g,t}-\sum_{l\in i}pl_{l,t}-Gs(2vb_{i,t}-u_{i,t})\\ & \sum_{b\in i}q_{b,t}=\sum_{g\in i}qg_{g,t}-\sum_{l\in i}ql_{l,t}+Bs(2vb_{i,t}-u_{i,t}) \end{align*}\]

Generator Startup Models

use @autodocs block to splice the docstrings

EGRIP.form_bs_logic — Method

generator cranking constraint

Once a non-black start generator is on, that is, $y_{g,t}=1$, then it needs to absorb the cranking power for its corresponding cranking time

"After" the time step that this unit satisfies its cranking constraint, its power goes to zero; and from the next time step, it becomes a dispatchable generator

set non-black start unit generation limits based on "generator cranking constraint"

cranking constraint states if generator g has absorb the cranking power for its corresponding cranking time, it can produce power

Mathematically if there exist enough 1 for $y_{g,t}=1$, then enable this generator's generating capability

There will be the following scenarios

(1) generator is off, then $y_{g,t}-y_{g,t-Tcr_{g}} = 0$, then $pg_{g,t} = 0$

(2) generator is on but cranking time not satisfied, then $y_{g,t} - y_{g,t-Tcr_g} = 1$, then $pg_{g,t} = -Pcr_g$

(3) generator is on and just satisfies the cranking time, then $y_{g,t} - y_{g,t-Tcr_g} = 0$, $y_{g,t-Tcr_g-1}=0$, then $pg_{g,t} = 0$

(4) generator is on and bigger than satisfies the cranking time, then $y_{g,t} - y_{g,t-Tcr_g} = 0$, $y_{g,t-Tcr_g-1}=1$, then $0 <= pg_{g,t} <= pg^{\max}_{g}$

All scenarios can be formulated as follows:

\[\begin{align*} & pg^{\min}_{g} \leq pg_{g,t} \leq pg^{\max}_{g}\\ & \text{ if }t > Tcr_{g}+1\\ & \quad\quad -Pcr_{g}(y_{g,t}-y_{g,Tcr_{g}}) \leq pg_{g,t} \leq pg^{\max}_{g}y_{g,t-Tcr_{g}-1}-Pcr_{g}(y_{g,t} - y_{g,t-Tcr_{g}}) \\ & \text{ elseif }t \leq Tcr_{g}\\ & \quad\quad pg_{g,t} = -Pcr_{g}y_{g,t}\\ & \text{else }\\ & \quad\quad pg_{g,t} = -Pcr_{g}(y_{g,t} - y_{g,1}) \end{align*}\]

EGRIP.form_gen_logic — Method

generator status and output constraint

generator ramping rate constraint

\[\begin{align*} -Krp_{g} \leq pg_{g,t}-pg_{g,t+1} \leq Krp_{g} \end{align*}\]

black-start unit is determined by the cranking power

\[\begin{align*} y_{g,t}=1 \text{ if } Pcr_{g}=0 \end{align*}\]

on-line generators cannot be shut down

\[\begin{align*} y_{g,t} <= y_{g,t+1} \end{align*}\]

EGRIP.form_load_logic — Method

load pickup constraint

restored load cannot exceed its maximum values

\[\begin{align*} & 0 \leq pl_{l,t} \leq pl^{\max}u_{l,t}\\ & 0 \leq ql_{l,t} \leq ql^{\max}u_{l,t}\\ \end{align*}\]

restored load cannot be shed

\[\begin{align*} & pl_{l,t-1} \leq pl_{l,t}\\ & ql_{l,t-1} \leq ql_{l,t}\\ \end{align*}\]

<!– ## I/O Function use @autodocs block to splice the docstrings

–>