Internal Library

Formulation Function

use @autodocs block to splice the docstrings

EGRIP.solve_restoration — Method

Solve restoration problem including the following constraints:

linearized AC power flow constraint

steady-state voltage variation constraint

generator cranking constraint

generator status and output constraint

load pick-up constraint

I/O Function

use @autodocs block to splice the docstrings

EGRIP._check_conductors — Method

EGRIP._check_cost_function — Method

EGRIP._make_mixed_units — Method

EGRIP._make_per_unit — Method

EGRIP._rescale_cost_model — Method

EGRIP._simplify_pwl_cost — Function

checks the slope of each segment in a pwl function, simplifies the function if the slope changes is below a tolerance

EGRIP.add_dcline_costs — Method

adds dcline costs, if gen costs exist

EGRIP.add_line_delimiter — Method

EGRIP.apply_func — Method

EGRIP.arrays_to_dicts! — Method

turns top level arrays into dicts

EGRIP.check_branch_directions — Method

checks that all parallel branches have the same orientation

EGRIP.check_branch_loops — Method

checks that all branches connect two distinct buses

EGRIP.check_bus_types — Method

checks bus types are consistent with generator connections, if not, fixes them

EGRIP.check_conductors — Method

EGRIP.check_connectivity — Method

checks that all buses are unique and other components link to valid buses

EGRIP.check_cost_functions — Method

throws warnings if cost functions are malformed

EGRIP.check_current_limits — Method

checks that each branch has a reasonable current rating-a, if not computes one

EGRIP.check_dcline_limits — Method

checks that parameters for dc lines are reasonable

EGRIP.check_keys — Method

EGRIP.check_network_data — Method

Runs various data quality checks on a PowerModels data dictionary. Applies modifications in some cases. Reports modified component ids.

EGRIP.check_storage_parameters — Method

checks that each storage unit has a reasonable parameters

EGRIP.check_thermal_limits — Method

checks that each branch has a reasonable thermal rating-a, if not computes one

EGRIP.check_transformer_parameters — Method

checks that each branch has a reasonable transformer parameters this is important because setting tap == 0.0 leads to NaN computations, which are hard to debug

EGRIP.check_type — Method

Checks if the given value is of a given type, if not tries to make it that type

EGRIP.check_voltage_angle_differences — Function

checks that voltage angle differences are within 90 deg., if not tightens

EGRIP.check_voltage_setpoints — Method

throws warnings if generator and dc line voltage setpoints are not consistent with the bus voltage setpoint

EGRIP.make_mixed_units — Method

Transforms network data into mixed-units (inverse of per-unit)

EGRIP.make_per_unit — Method

Transforms network data into per-unit

EGRIP.matpower_to_powermodels — Method

Converts a Matpower dict into a PowerModels dict

EGRIP.merge_bus_name_data — Method

merges bus name data into buses, if names exist

EGRIP.merge_generator_cost_data — Method

merges generator cost functions into generator data, if costs exist

EGRIP.merge_generic_data — Method

merges Matpower tables based on the table extension syntax

EGRIP.mp2pm_branch — Method

sets all branch transformer taps to 1.0, to simplify branch models

EGRIP.mp2pm_dcline — Method

adds pmin and pmax values at to and from buses

EGRIP.mp_cost_data — Method

EGRIP.parse_matlab_cells — Method

EGRIP.parse_matlab_data — Method

EGRIP.parse_matlab_matrix — Method

EGRIP.parse_matpower_file — Method

EGRIP.parse_matpower_file — Method

EGRIP.parse_matpower_string — Method

EGRIP.parse_mfile — Method

Parses the matpower data from either a filename or an IO object

EGRIP.row_to_dict — Method

takes a row from a matrix and assigns the values names

EGRIP.simplify_cost_terms — Method

trims zeros from higher order cost terms

EGRIP.split_line — Method

EGRIP.split_loads_shunts — Method

splitloadsshunts(data) Seperates Loads and Shunts in data under separate "load" and "shunt" keys in the PowerModels data format. Includes references to originating bus via "loadbus" and "shuntbus" keys, respectively.

EGRIP.type_array — Method

Attempts to determine the type of an array of strings extracted from a matlab file

EGRIP.type_value — Method

Attempts to determine the type of a string extracted from a matlab file