Public Library

Formulation Function

EGRIP.solve_restoration_full — Function

Solve full restoration problem (The restoration problem could be partial or full restorations)

Full restoration problem assumes that the entire network is broken down.

Inputs: A set of restoration data in csv format and original system data

network data directory where data format could be json, matpower and psse

restoration data directory

restoration_gen: specify initial generator status, cranking specifications and black-start generators

restoration_bus: specify initial bus status and its load priority, from where the problem type (partial or full restorations) can be determined

restoration_line: specify initial line status, from where the problem type (partial or full restorations) can be determined

result storage directory

gap

Output: Restoration plans

Constraints:

linearized AC power flow constraint

steady-state voltage variation constraint

generator cranking constraint

generator status and output constraint

load pick-up constraint

Missing docstring.

Missing docstring for solve_restoration_part. Check Documenter's build log for details.

EGRIP.solve_section — Function

Solve sectionalization problem for restoration preparedness

Problem type: The sectionalization problem could assign all buses to certain sections or only critical buses.

Inputs:

network data directory

restoration data directory

result storage directory

gap

Output:

JSON file of network data of each section

Constraints:

EGRIP.solve_startup — Function

Solve generator start-up problem

Inputs: A set of restoration data in csv format and original system data

network data directory where data format could be json, matpower and psse

restoration data directory

restoration_gen: specify initial generator status, cranking specifications and black-start generators

restoration_bus: specify initial bus status and its load priority, from where the problem type (partial or full restorations) can be determined

restoration_line: specify initial line status, from where the problem type (partial or full restorations) can be determined

result storage directory

gap

formulation types

specifications of wind power

wind activation 0 denotes no wind will support the restoration

wind activation 1 denotes wind will support the restoration with its distribution as normal

wind activation 2 denotes wind will support the restoration with time seires wind data

wind activation 3 denotes wind will support the restoration with real quantile and interpolation (linear, cubic, or spline)

if wind activation = 3, then real wind power density will be specified in wind_data

The followings are the keyword arguments

wind_data: time series wind data

saa_mode: different chance constraints enforcement in sample average approximations

Output: Generator start-up sequence

Constraints:

generato cranking

load pickup

power balance