Internal Library

Integrated Restoration

EGRIP.solve_restoration_full — Function

Solve full restoration problem (The restoration problem could be partial or full restorations)

Full restoration problem assumes that the entire network is broken down.

Inputs: A set of restoration data in csv format and original system data

network data directory where data format could be json, matpower and psse

restoration data directory

restoration_gen: specify initial generator status, cranking specifications and black-start generators

restoration_bus: specify initial bus status and its load priority, from where the problem type (partial or full restorations) can be determined

restoration_line: specify initial line status, from where the problem type (partial or full restorations) can be determined

result storage directory

gap

Output: Restoration plans

Constraints:

linearized AC power flow constraint

steady-state voltage variation constraint

generator cranking constraint

generator status and output constraint

load pick-up constraint

EGRIP.solve_restoration_part — Method

Solve partial restoration problem (The restoration problem could be partial or full restorations)

Partial restoration problem assumes that a part of the network is still functioning.

Inputs: A set of restoration data in csv format and original system data

network data directory where data format could be json, matpower and psse

restoration data directory

restoration_gen: specify initial generator status, cranking specifications and black-start generators

restoration_bus: specify initial bus status and its load priority, from where the problem type (partial or full restorations) can be determined

restoration_line: specify initial line status, from where the problem type (partial or full restorations) can be determined

result storage directory

gap

Output: Restoration plans

Constraints:

linearized AC power flow constraint

steady-state voltage variation constraint

generator cranking constraint

generator status and output constraint

load pick-up constraint

Sectionalization Models

EGRIP.solve_section — Method

Solve sectionalization problem for restoration preparedness

Problem type: The sectionalization problem could assign all buses to certain sections or only critical buses.

Inputs:

network data directory

restoration data directory

result storage directory

gap

Output:

JSON file of network data of each section

Constraints:

Generator Startup Models

EGRIP.solve_startup — Function

Solve generator start-up problem

Inputs: A set of restoration data in csv format and original system data

network data directory where data format could be json, matpower and psse

restoration data directory

restoration_gen: specify initial generator status, cranking specifications and black-start generators

restoration_bus: specify initial bus status and its load priority, from where the problem type (partial or full restorations) can be determined

restoration_line: specify initial line status, from where the problem type (partial or full restorations) can be determined

result storage directory

gap

formulation types

specifications of wind power

wind activation 0 denotes no wind will support the restoration

wind activation 1 denotes wind will support the restoration with its distribution as normal

wind activation 2 denotes wind will support the restoration with time seires wind data

wind activation 3 denotes wind will support the restoration with real quantile and interpolation (linear, cubic, or spline)

if wind activation = 3, then real wind power density will be specified in wind_data

The followings are the keyword arguments

wind_data: time series wind data

saa_mode: different chance constraints enforcement in sample average approximations

Output: Generator start-up sequence

Constraints:

generato cranking

load pickup

power balance

Generator Models

EGRIP.def_var_gen — Method

Define generator variables

The definitation of variables corresponds to different formulations.

EGRIP.form_gen_cranking_1 — Method

Generator cranking constraint (formulation 1)

Once a non-black start generator is on, that is, $y_{g,t}=1$, then it needs to absorb the cranking power for its corresponding cranking time

"After" the time step that this unit satisfies its cranking constraint, its power goes to zero; and from the next time step, it becomes a dispatchable generator

set non-black start unit generation limits based on "generator cranking constraint"

cranking constraint states if generator g has absorb the cranking power for its corresponding cranking time, it can produce power

Mathematically if there exist enough 1 for $y_{g,t}=1$, then enable this generator's generating capability

There will be the following scenarios

(1) generator is off, then $y_{g,t}-y_{g,t-Tcr_{g}} = 0$, then $pg_{g,t} = 0$

(2) generator is on but cranking time not satisfied, then $y_{g,t} - y_{g,t-Tcr_g} = 1$, then $pg_{g,t} = -Pcr_g$

(3) generator is on and just satisfies the cranking time, then $y_{g,t} - y_{g,t-Tcr_g} = 0$, $y_{g,t-Tcr_g-1}=0$, then $pg_{g,t} = 0$

(4) generator is on and bigger than satisfies the cranking time, then $y_{g,t} - y_{g,t-Tcr_g} = 0$, $y_{g,t-Tcr_g-1}=1$, then $0 <= pg_{g,t} <= pg^{\max}_{g}$

All scenarios can be formulated as follows:

\[\begin{align*} & pg^{\min}_{g} \leq pg_{g,t} \leq pg^{\max}_{g}\\ & \text{ if }t > Tcr_{g}+1\\ & \quad\quad -Pcr_{g}(y_{g,t}-y_{g,t-Tcr_{g}}) \leq pg_{g,t} \leq pg^{\max}_{g}y_{g,t-Tcr_{g}-1}-Pcr_{g}(y_{g,t} - y_{g,t-Tcr_{g}}) \\ & \text{ elseif }t \leq Tcr_{g}\\ & \quad\quad pg_{g,t} = -Pcr_{g}y_{g,t}\\ & \text{else }\\ & \quad\quad pg_{g,t} = -Pcr_{g}(y_{g,t} - y_{g,1}) \end{align*}\]

EGRIP.form_gen_cranking_2 — Method

Generator cranking constraint (formulation 2)

In the generator start-up sequence optimization problem, we consider the following cranking procedure

\[\begin{align*} p_{g}(x_g,t)=\begin{cases} 0 & 0\le t<x_{g}\\ -c_{g} & x_{g}\le t < x_{g}+t_{g}^{c}\\ r_g(t-x_{g}-t_{g}^{c}) & x_{g}+t_{g}^{c}\leq t < x_{g}+t_{g}^{c}+t_{g}^{r}\\ p^{\mbox{max}} & x_{g}+t_{g}^{c}+t_{g}^{r}\leq t \leq T, \end{cases} \end{align*}\]

We introduce binary variables $y_{gt}$, $z_{gt}$, and $u_{gt}$ to indicate whether generator $g$ is in status of cranking, ramping, or full capacity in time period $t$, respectively. We introduce binary variable $x_{gt}$ to indicate whether generator $g$ is started in time period $t$. The formulations are implemented in function form_gen_cranking_2.

First, a NBS generator has no activity before it is started

\[\begin{align*} &\begin{aligned} (t-1)(1-x_{gt}) \ge \sum_{i=1}^{t-1}y_{gi}\quad\forall g\in G, t\in T\backslash\{1\} \end{aligned}\\ &\begin{aligned} (t+t^c_g-1)(1-x_{gt}) \ge \sum_{i=1}^{f(t)} z_{gi}\quad\forall g\in G, t\in T \end{aligned}\\ &\begin{aligned} (t+t^c_g+t^r_g-1)(1-x_{gt}) \ge \sum_{i=1}^{g(t)}u_{gi}\quad\forall g\in G, t\in T \end{aligned}\\ & f(t)=\min\{|T|,t+t^c_g-1\},g(t)=\min\{|T|,t+t^c_g +t^r_g-1\} \end{align*}\]

Second, once started, a cranking is followed

\[\begin{align*} \sum_{i=t}^{f(t)} y_{gi}\ge x_{gt}\times\min\{|T|-t, t^c_g\}\quad \forall t\in T \end{align*}\]

Third, once cranking is finished, a ramping is followed

\[\begin{align*} \label{eq_gen_ramp} \sum_{i=t+t^r_g-1}^{g(t)} z_{gi}\ge x_{gt}\times\min\{|T|-t, t^r_g\}\quad \forall t\in T \end{align*}\]

Once the generator reaches to its maximum value, it should stay in this status

\[\begin{align*} \label{eq_gen_max} u_{gt}\geq u_{gt-1}\quad \forall t\in T\backslash\{1\} \end{align*}\]

Additionally, in each stage each generator will have one status being activated, i.e.,

\[\begin{align*} \label{eq_gen_one_status} y_{gi} + z_{gi} + u_{gi}\leq 1\quad \forall t\in T \end{align*}\]

And there exists only one validated generator start-up moment

\[\begin{align*} \label{eq_gen_one_start} \sum_{t\in T}x_{gt} =1\quad \forall g\in G \end{align*}\]

We can write the generation output of unit $g$ in time period $t$ as

\[\begin{align*} \label{eq_gen_power} p_g(t) = -c_gy_{gt}+ \sum_{i=1}^t{z_{gi}r_g} \end{align*}\]

EGRIP.form_gen_cranking_3 — Method

Generator cranking constraint (formulation 3)

In the generator start-up sequence optimization problem, we consider the following cranking procedure

\[\begin{align*} p_{g}(x_g,t)=\begin{cases} 0 & 0\le t<x_{g}\\ -c_{g} & x_{g}\le t < x_{g}+t_{g}^{c}\\ r_g(t-x_{g}-t_{g}^{c}) & x_{g}+t_{g}^{c}\leq t < x_{g}+t_{g}^{c}+t_{g}^{r}\\ p^{\mbox{max}} & x_{g}+t_{g}^{c}+t_{g}^{r}\leq t \leq T, \end{cases} \end{align*}\]

If $t-x_{g}<0$, $p_{gt}=0$. Introduce positive large number $M$ and binary variable $a_{gt}$. Build $t - x_{g} < 0 \Leftrightarrow a_{gt}=1$ and $t - x_{g} \geq 0 \Leftrightarrow a_{gt}=0$, where $a_{gt}=1$ indicates generator $g$ is off-line.

\[ \begin{align*} \begin{aligned} &t - x_{g}\leq M (1- a_{gt}) \\ &t-x_g \geq -M a_{gt} \\ &-M (1- a_{gt}) \leq p_{gt}\\ &p_{gt}\leq M (1- a_{gt}) \end{aligned} \end{align*}\]

If $x_{g}\le t $ and $t < x_{g}+t_{g}^{c}$, $p_{gt}=-c_{g}$. Introduce positive large number $M$ and binary variable $b_{gt}$. Build $t < x_{g}+t_{g}^{c} \Leftrightarrow b_{gt}=1$ and $t \geq x_{g}+t_{g}^{c} \Leftrightarrow b_{gt}=0$, where $b_{gt}=1$ indicates generator $g$ is at the cranking stage.

\[\begin{align*} \begin{aligned} &t - x_{g}- t_{g}^{c} \leq M (1- b_{gt}) \\ &t - x_{g}- t_{g}^{c}\geq -M b_{gt} \\ &-M (1+a_{gt}- b_{gt}) \leq p_{gt} + c_{g}\\ & p_{gt} + c_{g}\leq M (1+a_{gt}- b_{gt}) \end{aligned} \end{align*}\]

If $t < x_{g}+t_{g}^{c}+t_{g}^{r}$, $p_{gt}=r_g(t-x_{g}-t_{g}^{c})$. Introduce positive large number $M$ and binary variable $c_{gt}$. Build $t < x_{g}+t_{g}^{c}+t_{g}^{r} \Leftrightarrow c_{gt}=1$ and $t \geq x_{g}+t_{g}^{c}+t_{g}^{r} \Leftrightarrow c_{gt}=0$, where $c_{gt}=1$ indicates generator $g$ is at the ramping stage.

\[\begin{align*} \begin{aligned} &t - (x_{g}+t_{g}^{c}+t_{g}^{r}) \leq M (1- c_{gt}) \\ &t - (x_{g}+t_{g}^{c}+t_{g}^{r})\geq -M c_{gt} \\ &-M (1+a_{gt}+ b_{gt}- c_{gt}) \leq p_{gt} - r_g(t-x_{g}-t_{g}^{c})\\ & p_{gt} - r_g(t-x_{g}-t_{g}^{c})\leq M (1+a_{gt}+ b_{gt}- c_{gt}) \end{aligned} \end{align*}\]

For the last stage, we do not need to introduce new variables.

\[\begin{align*} &-M (a_{gt}+ b_{gt} + c_{gt}) \leq p_{gt} - p^{\mbox{max}} \\ & p_{gt} - p^{\mbox{max}} \leq M (a_{gt}+ b_{gt} + c_{gt}) \end{align*}\]

EGRIP.form_gen_logic — Method

generator status and output constraint

generator ramping rate constraint

\[\begin{align*} -Krp_{g} \leq pg_{g,t}-pg_{g,t+1} \leq Krp_{g} \end{align*}\]

black-start unit is determined by the cranking power

\[\begin{align*} y_{g,t}=1 \text{ if } Pcr_{g}=0 \end{align*}\]

on-line generators cannot be shut down

\[\begin{align*} y_{g,t} <= y_{g,t+1} \end{align*}\]

Load Models

EGRIP.def_var_load — Method

Define load variables

EGRIP.form_load_logic — Method

Load pickup constraint used for full restoration problem

restored load cannot exceed its maximum values

\[\begin{align*} & 0 \leq pl_{l,t} \leq pl^{\max}u_{l,t}\\ & 0 \leq ql_{l,t} \leq ql^{\max}u_{l,t}\\ \end{align*}\]

restored load cannot be shed

\[\begin{align*} & pl_{l,t-1} \leq pl_{l,t}\\ & ql_{l,t-1} \leq ql_{l,t}\\ \end{align*}\]

EGRIP.form_load_logic_1 — Method

Load pickup constraint form 1 used in generator start-up problem

EGRIP.form_load_logic_2 — Method

Load pickup constraint (formulation 2)

The load sequential actions are described below:

\[\begin{align*} p_{d}(x_d,t)=\begin{cases} 0 & 0\le t < x_{d}\\ -p_{d} & x_{d}\le t\le T, \end{cases} \end{align*}\]

We introduce $y_{dt}$ to indicate whether load $d$ is in status of being energized and $x_{dt}$ to indicate whether the energization event of load $d$ happens in time $t$. The formulations are implemented in form_load_logic_2.

First, a load has no activity before it is picked up

\[\begin{align*} \label{eq_load_no_action} \sum_{i=1}^{t-1}y_{di} \le (t-1)(1-x_{dt})\quad \forall d\in D,t\in T\backslash\{1\} \end{align*}\]

Second, a load is served to the end of the time horizon once it is picked up

\[\begin{align*} \label{eq_load_served} \sum_{i=t}^{|T|}y_{di} \ge (|T|-t+1)x_{dt}\quad \forall d\in D \end{align*}\]

And there exists only one validated load energization moment

\[\begin{align*} \label{eq_load_one_start} \sum_{t\in T}x_{gt} =1\quad \forall g\in G \end{align*}\]

We can write the load $d$ in time period $t$ as

\[\begin{align*} \label{eq_load_power} p_d(t) = - y_{dt}p_d \end{align*}\]

EGRIP.form_load_logic_3 — Method

Load pickup constraint (formulation 3)

The load sequential actions are described below:

\[\begin{align*} p_{d}(x_d,t)=\begin{cases} 0 & 0\le t < x_{d}\\ -p_{d} & x_{d}\le t\le T, \end{cases} \end{align*}\]

If $t-x_{d}<0$, $p_{dt}=0$. Introduce positive large number $M$ and binary variable $e_{dt}$. Build $t-x_{d}<0\Leftrightarrow e_{dt}=1$ and $t-x_{d}\geq 0 \Leftrightarrow e_{dt}=0$, where $e_{dt}=1$ indicates load $d$ is off-line.

\[\begin{align*} \begin{aligned} &t - x_{d} \leq M (1- e_{dt}) \\ &t - x_{d} \geq -M e_{dt} \\ &0 \leq p_{dt} \leq M (1- e_{dt})\\ &-M e_{dt} \leq p_{dt} + p_{d} \leq M e_{dt}\\ \end{aligned} \end{align*}\]

Power Flow Models

EGRIP.def_var_flow — Method

Define flow variable

EGRIP.form_branch — Method

Branch (power flow) constraints

linearized power flow

\[\begin{align*} p_{bij,t}=G_{ii}(2vl_{ij,t}-x_{ij,t}) + G_{ij}(vl_{ij,t} + vl_{ji,t}-x_{ij,t}) + B_{ij}(al_{ij,t}-al_{ij,t})\\ q_{bij,t}=-B_{ii}(2vl_{ij,t}-x_{ij,t}) - B_{ij}(vl_{ij,t} + vl_{ji,t}-x_{ij,t}) + G_{ij}(al_{ij,t}-al_{ij,t})\\ \end{align*}\]

EGRIP.form_nodal — Method

Form the nodal constraints:

voltage constraint

voltage deviation should be limited

voltage constraints are only activated if the associated line is energized

\[\begin{align*} & v^{\min}_{i} \leq v_{i,t} \leq v^{\max}_{i}\\ & v^{\min}_{i}x_{ij,t} \leq vl_{ij,t} \leq v^{\max}_{i}x_{ij,t}\\ & v^{\min}_{j}x_{ij,t} \leq vl_{ji,t} \leq v^{\max}_{j}x_{ij,t}\\ & v_{i,t} - v^{\max}_{i}(1-x_{ij,t}) \leq vl_{ij,t} \leq v_{i,t} - v^{\min}_{i}(1-x_{ij,t})\\ & v_{j,t} - v^{\max}_{j}(1-x_{ij,t}) \leq vl_{ij,t} \leq v_{j,t} - v^{\min}_{j}(1-x_{ij,t}) \end{align*}\]

angle difference constraint

angle difference should be limited

angle difference constraints are only activated if the associated line is energized

\[ \begin{align*} & a^{\min}_{ij} \leq a_{i,t}-a_{j,t} \leq a^{\max}_{ij}\\ & a^{\min}_{ij}x_{ij,t} \leq al_{ij,t}-al_{ji,t} \leq a^{\max}_{ij}x_{ij,t}\\ & a_{i,t}-a_{j,t}-a^{\max}_{ij}(1-x_{ij,t}) \leq al_{ij,t}-al_{ji,t} \leq a_{i,t}-a_{j,t}-a^{\min}_{ij}(1-x_{ij,t}) \end{align*}\]

generator and bus energizing logics

energized line cannot be shut down

bus should be energized before the connected genertor being on

\[\begin{align*} & x_{ij,t} \geq x_{ij,t-1}\\ & u_{i,t} \geq x_{ij,t}\\ & u_{j,t} \geq x_{ij,t} \end{align*}\]

bus energized constraints

bus energized indicating generator energized

energized buses cannot be shut down

\[\begin{align*} & v^{\min}u_{i,t} \leq vb_{i,t} \leq v^{\max}u_{i,t} \\ & v_{i,t} - v^{\max}(1-u_{i,t}) \leq vb_{i,t} \leq v_{i,t} - v^{\min}(1-u_{i,t})\\ & u_{g,t} = y_{g,t}\\ & u_{i,t} \geq u_{i,t-1} \end{align*}\]

nodal power balance constraint

\[\begin{align*} & \sum_{b\in i}p_{b,t}=\sum_{g\in i}pg_{g,t}-\sum_{l\in i}pl_{l,t}-Gs(2vb_{i,t}-u_{i,t})\\ & \sum_{b\in i}q_{b,t}=\sum_{g\in i}qg_{g,t}-\sum_{l\in i}ql_{l,t}+Bs(2vb_{i,t}-u_{i,t}) \end{align*}\]

Renewable Models

EGRIP.def_var_wind — Method

Define wind generator variables

EGRIP.form_wind_saa_1 — Method

Form wind power dispatch chance constraints approximated by Sample Averaged Approximation This function is associated with wind activation 1 option: wind data from normal distribution

EGRIP.form_wind_saa_2 — Method

Form wind power dispatch chance constraints approximated by Sample Averaged Approximation This function is associated with wind activation 2 option: wind data from real time series in a Dictionary

EGRIP.form_wind_saa_3 — Method

Form wind power dispatch chance constraints approximated by Sample Averaged Approximation This function is associated with wind activation 3 option: wind data given by an estimated density The chance constraints in sample-average approximation can be enforced into samples through all periods or individual periods

Mode 1: The chance constraints can be enforced into samples through all periods

Mode 2: The chance constraints can be enforced into samples through individual periods

Util Functions

EGRIP.load_gen — Method

Load generator data with respect to restoration

EGRIP.load_network — Method

Load network data